+ switch to refactored render.set_sampling operator
This shall make AMI setup a lot easier, as only a few scripts need to be cloned into the startup folder from now on. Time to think about starting to envision the possiblity to create an install.sh file on the AWS image. It should basically update the currentDev build, as well as clone the latest git repo to the scripts/startup folder. That way it will not be necessary any longer to clone the full 1GB repo into the AMI.